Updated Settings for Default Values
So I decided to go through every single setting and supply a default
value to each one. If everyone uses this as standard in the future there
should be no need to switch out configs, it will simply update the old
config with the new default values.

IF THIS WORKS, MORE WORK WILL BE NECESSARY TO REMOVE OLD CONFIG
REPLACEMENT.