Added Translation Serialization Error Handling
Changed so that if an error occurs in the loading of a translation file.

This often occurred if a type was made in the Key of a file. Currently
it can only resolve issues with the EN file however will display issues
with alternate languages.

I've also created a ThrowFatalError method which I quite like.